+++ app/justfile
@@ -82,12 +82,12 @@ status-test queue="local1":
 status-prod queue="public1":
     curl https://container.mtfm.io/{{ queue }}/status | jq .
 
-test: check
+test mode="remote" +args="": (_validate_mode mode)
     #!/usr/bin/env bash
     set -e # exit when any command fails, preserving the status of the failed command
 
     cleanup() {
-        docker compose --project-name=test -f docker-compose.yml -f docker-compose-remote.yml down 2>/dev/null || true
+        docker compose --project-name=test -f docker-compose.yml -f docker-compose-{{ mode }}.yml down 2>/dev/null || true
     }
 
     trap cleanup EXIT # cleanup when the script exits for any reason, including failure
@@ -98,8 +98,6 @@ test: check
     get_free_port() {
         while true; do
             local port=$(shuf -i 10000-65535 -n 1)
-            # Check if port is in use using netstat or ss
-            # Check if port is in use using netstat/ss on Linux or lsof on macOS
             if [ "$(uname)" == "Linux" ]; then
                 if ! ss -tuln | grep -q ":${port} "; then
                     echo "${port}"
@@ -116,6 +114,16 @@ test: check
 
     export APP_PORT=$(get_free_port)
     echo -e "Using port ${APP_PORT}"
+
+    # Set API_URL and QUEUE_ID based on mode
+    if [ "{{ mode }}" = "local" ]; then
+        export API_URL="http://worker:8000"
+        export QUEUE_ID="local"
+    else
+        export API_URL="http://api1:8081"
+        export QUEUE_ID="local1"
+    fi
+
     # Get host IP address for Linux
     if [ "$(uname)" == "Linux" ]; then
         HOST_IP=$(ip -4 addr show docker0 | grep -Po 'inet \K[\d.]+')
@@ -123,15 +131,15 @@ test: check
         HOST_IP="host.docker.internal"
     fi
 
-    # A local NPM registry saves time, bandwidth, and avoids rate limits
-    # This fails on specific npm modules, so we'll just not use it for now
-    # STATUS_CODE=$(curl --write-out '%{http_code}' --silent --output /dev/null http://localhost:4873 || true)
-    # if [ $STATUS_CODE = "200" ]; then
-    #     export NPM_CONFIG_REGISTRY=http://${HOST_IP}:4873
-    # fi
-    # This is causing currently unknown issues, a dangling worker 
+    export AWS_SECRET_ACCESS_KEY="minioaccesskey123123"
+    export AWS_ACCESS_KEY_ID="minioaccesskey123123"
+    export AWS_ENDPOINT="https://minio.worker-metaframe.localhost"
+    export AWS_REGION="us-west-2"
+    export AWS_S3_BUCKET="metaframe-asman-test"
+    export REDIS_URL="redis://localhost:6379" 
+
     docker rm test-worker-1 2>/dev/null || true
-    docker compose --project-name=test -f docker-compose.yml -f docker-compose-remote.yml up --build --remove-orphans --detach
+    docker compose --project-name=test -f docker-compose.yml -f docker-compose-{{ mode }}.yml up --build --remove-orphans --detach
     just worker/test
     just test/test
     just shared/test
