+++ app/shared/src/mod.ts
@@ -3,3 +3,5 @@ export * from "./shared/dataref.ts";
 export * from "./shared/jobtools.ts";
 export * from "./shared/types.ts";
 export * from "./shared/util.ts";
+export * from "./shared/jobqueue.ts";
+export * from "./shared/db.ts";
