+++ app/worker/assets
@@ -1,0 +1,1 @@
+../browser/dist
